trading/requirements.txt
websockets>=14.0
httpx>=0.27
apscheduler>=3.10
python-dotenv>=1.0
pytz>=2024.1
pydantic>=2.0
beautifulsoup4>=4.12
feedparser>=6.0
